Fix duplicate Love2D startup by making root entrypoints authoritative (#18)
* Inline Love2D entrypoints in root files

* Remove redundant window reset on startup
